refactor(logic): isolate canonical CNF contracts
